Refactor in order to add tests for week calculations
