+++ Gemfile
@@ -2,3 +2,6 @@ source 'https://rubygems.org'
 
 gem 'sinatra'
 gem 'activesupport'
+
+gem 'rack-test'
+gem 'rspec'
