+++ spec/spec_helper.rb
@@ -1,0 +1,13 @@
+require 'rack/test'
+require 'rspec'
+
+ENV['RACK_ENV'] = 'test'
+
+require File.expand_path '../../application.rb', __FILE__
+
+module RSpecMixin
+  include Rack::Test::Methods
+  def app() described_class end
+end
+
+RSpec.configure { |c| c.include RSpecMixin }
